Changed all int to Integer and long to Long and fixed POST createPlayer to return created player with ID.

+++ spring_jdbc/src/main/java/com/packt/footballpg/controllers/TeamsController.java
@@ -30,7 +30,7 @@ public class TeamsController {
   }
 
   @GetMapping("/{id}")
-  public Team getTeamById(@PathVariable int id) {
+  public Team getTeamById(@PathVariable Long id) {
     return teamsService.getTeamById(id);
   }
 }
