+++ spring_jdbc/src/main/java/com/packt/footballpg/records/Player.java
@@ -2,5 +2,5 @@ package com.packt.footballpg.records;
 
 import java.time.LocalDate;
 
-public record Player(Integer id, Integer jerseyNumber, String name, String position, LocalDate dateOfBirth, Integer teamId) {
+public record Player(Long id, Integer jerseyNumber, String name, String position, LocalDate dateOfBirth, Long teamId) {
 }
